components/model_upload_cicd_trigger/requirements.txt
google-cloud-storage==2.10.0
requests==2.31.0